Config moved to separate folder
