+++ config/default.js
@@ -1,0 +1,8 @@
+'use strict';
+
+var config = {};
+
+config.port = 8080;
+config.mongouri = 'mongodb://localhost:27017/expenseapptest';
+
+module.exports = config;
